@@ -1,9 +1,7 @@
-import typing as _t
 import sys
 import asyncio
 import signal
 import socket
-import time
 
 import uvicorn
 import uwsgi  # pylint: disable=import-error
@@ -12,11 +10,21 @@ from django.conf import settings
 from django.utils.module_loading import import_string
 
 try:
-    import uvloop
-    uwsgi.log('uvloop installed.')
+    if not settings.RUN_WORKER:
+        import uvloop
+        uwsgi.log(f'uvloop installed on worker {uwsgi.worker_id()}.')
+    else:
+        uvloop = None
 except ImportError:
-    uwsgi.log('No uvloop installed. Run with native asyncio.')
+    uwsgi.log(f'No uvloop installed on worker {uwsgi.worker_id()}. Run with native asyncio.')
     uvloop = None
+
+try:
+    import h11
+    uwsgi.log(f'h11 installed on worker {uwsgi.worker_id()}.')
+except ImportError:
+    uwsgi.log(f'No h11 installed on worker {uwsgi.worker_id()}.')
+    h11 = None
 
 django.setup(set_prefix=False)
 
@@ -24,25 +32,36 @@ django.setup(set_prefix=False)
 application = import_string(settings.ASGI_APPLICATION)
 
 
-async def init(servers_list, fds, app):
+async def init(fds, app):
+    uvicorn_settings = settings.CONFIG['uvicorn'].all()
+    if uvloop:
+        uvicorn_settings.setdefault('loop', 'uvloop')
+    if h11:
+        uvicorn_settings.setdefault('http', 'h11')
     config = uvicorn.Config(
         app=app,
         interface='asgi3',
         log_level=settings.LOG_LEVEL.lower(),
         headers=settings.WEB_SERVER_HEADERS or None,
-        **settings.CONFIG['uvicorn'].all(),
+        **uvicorn_settings,
     )
     server = uvicorn.Server(config)
-    servers_list.append(server)
+    servers.append(server)
+    # Spawn a handler for every uWSGI socket
     sockets = [
         socket.fromfd(fd, socket.AF_INET, socket.SOCK_STREAM)
         for fd in fds
     ]
     try:
+        # Run worker server
         await server.serve(sockets=sockets)
-        uwsgi.log(f'uvicorn server stopped after {server.server_state.total_requests} requests.')
+        uwsgi.log(
+            f'Server uvicorn stopped after {server.server_state.total_requests} requests on worker {uwsgi.worker_id()}.'
+        )
+    except KeyboardInterrupt:
+        uwsgi.log(f'Stopped uvicorn server on worker {uwsgi.worker_id()}')
     except BaseException as exc:
-        uwsgi.log(f"uvicorn say: {str(exc)}")
+        uwsgi.log(f"uvicorn say: {str(exc)!r} on worker {uwsgi.worker_id()}")
         raise
     finally:
         for sock in sockets:
@@ -51,42 +70,23 @@ async def init(servers_list, fds, app):
                 sock.close()
 
 
-def destroy():
-    uwsgi.log(f"destroy worker {uwsgi.worker_id()}")
+def handle_sigterm():
+    uwsgi.log(f'Catch SIGTERM signal on worker {uwsgi.worker_id()}')
     for server in servers:
         server.should_exit = True
-        server.handle_exit(sig=signal.SIGINT, frame=None)
-    time.sleep(1)
-    uwsgi.log(f"exit worker {uwsgi.worker_id()}")
-    sys.exit(0)
-
-
-def graceful_reload():
-    uwsgi.log(f"graceful reload for worker {uwsgi.worker_id()}")
-    for server in servers:
-        server.handle_exit(sig=signal.SIGTERM, frame=None)
-    time.sleep(1)
-    loop.stop()
-    loop.close()
-    sys.exit(0)
+        server.handle_exit(signal.SIGINT, None)
 
 
 if __name__ == '__main__':
     # Init event loop and destroy handlers
     loop = asyncio.new_event_loop() if uvloop is None else uvloop.new_event_loop()
-    loop.add_signal_handler(signal.SIGINT, destroy)
-    loop.add_signal_handler(signal.SIGTERM, destroy)
-    loop.add_signal_handler(signal.SIGHUP, graceful_reload)
+    loop.add_signal_handler(signal.SIGTERM, handle_sigterm)
 
-    # TODO: Does it nessesary?
-    servers: _t.List[uvicorn.Server] = []
+    uwsgi.log(f'Start uvicorn worker {uwsgi.worker_id()}')
+    servers: list[uvicorn.Server] = []
+    loop.run_until_complete(init(uwsgi.sockets, application))
 
-    # Spawn a handler for every uWSGI socket
-    uwsgi.log(f'start uvicorn worker {uwsgi.worker_id()}')
-    loop.run_until_complete(init(servers, uwsgi.sockets, application))
-
-    # Start accepting requests via socket
-    uwsgi.accepting()
-
-    # Graceful shutdown
-    destroy()
+    uwsgi.log(f'Exiting uvicorn worker {uwsgi.worker_id()}')
+    loop.stop()
+    loop.close()
+    sys.exit(0)
