@@ -1,2 +1,2 @@
 # pylint: disable=django-not-available
-__version__: str = '5.10.0b5'
+__version__: str = '5.10.0b6'
